@@ -6,6 +6,7 @@ package template
 
 import (
 	"github.com/gohugoio/hugo/tpl/internal/go_templates/texttemplate/parse"
+	"maps"
 	"reflect"
 	"sync"
 )
@@ -102,12 +103,8 @@ func (t *Template) Clone() (*Template, error) {
 	}
 	t.muFuncs.RLock()
 	defer t.muFuncs.RUnlock()
-	for k, v := range t.parseFuncs {
-		nt.parseFuncs[k] = v
-	}
-	for k, v := range t.execFuncs {
-		nt.execFuncs[k] = v
-	}
+	maps.Copy(nt.parseFuncs, t.parseFuncs)
+	maps.Copy(nt.execFuncs, t.execFuncs)
 	return nt, nil
 }
 
